+++ srcs/validation.c
@@ -32,201 +32,6 @@ static int  all_remaining(int i, char **arr)
     return (1);
 }
 
-// static int  check_inlets(char **map, size_t i, size_t j)
-// {
-//     int res;
-
-//     res = 0;
-//     if (i != 0 && i != len_2d(map) && j != 0 && j != ft_strlen(map[i]) - 1)
-//     {
-//         if (map[i - 1][j] != '1' && map[i - 1][j] != ' ') // what if the character is on the wall and a different char?
-//             res = 1;
-//         if (map[i + 1][j] != '1' && map[i + 1][j] != ' ')
-//             res = 2;
-//         if (map[i][j - 1] != '1' && map[i][j - 1] != ' ')
-//             res = 3;
-//         if (map[i][j + 1] != '1' && map[i][j + 1] != ' ')
-//             res = 4;
-//     }
-//     else if (i == 0 && j != 0 && j != ft_strlen(map[i]) - 1)
-//     {
-//         if (map[i + 1][j] != '1' && map[i + 1][j] != ' ')
-//             res = 5;
-//         if (map[i][j - 1] != '1' && map[i][j - 1] != ' ')
-//             res = 6;
-//         if (map[i][j + 1] != '1' && map[i][j + 1] != ' ')
-//             res = 7;
-//     }
-//     else if (i == len_2d(map) && j != 0 && j != ft_strlen(map[i]) - 1)
-//     {
-//         if (map[i - 1][j] != '1' && map[i - 1][j] != ' ')
-//             res = 8;
-//         if (map[i][j - 1] != '1' && map[i][j - 1] != ' ')
-//             res = 9;
-//         if (map[i][j + 1] != '1' && map[i][j + 1] != ' ')
-//             res = 10; 
-//     }
-//     else if (j == 0 && i != 0 && i != len_2d(map))
-//     {
-//         if (map[i - 1][j] != '1' && map[i - 1][j] != ' ')
-//             res = 11;
-//         if (map[i + 1][j] != '1' && map[i + 1][j] != ' ')
-//             res = 12;
-//         if (map[i][j + 1] != '1' && map[i][j + 1] != ' ')
-//             res = 13;
-//     }
-//     else if (j != ft_strlen(map[i]) - 1 && i != 0 && i != len_2d(map))
-//     {
-//         if (map[i - 1][j] != '1' && map[i - 1][j] != ' ')
-//             res = 14;
-//         if (map[i + 1][j] != '1' && map[i + 1][j] != ' ')
-//             res = 15;
-//         if (map[i][j - 1] != '1' && map[i][j - 1] != ' ')
-//             res = 16;
-//     }
-//     if (res > 0)
-//         print_error("Map not surrounded by walls");
-//     return (res);
-// }
-
-// static void check_walls(t_map *map)
-// {
-//     size_t i;
-//     size_t len;
-//     size_t j;
-    
-//     i = 0;
-//     len = len_2d(map->map);
-//     j = 0;
-//     while (map->map[i][j]) // first char in column
-//     {
-//         i = 0;
-//         while (map->map[i][j] == ' ')
-//         {
-//             check_inlets(map->map, i, j);
-//             i++;
-//         }
-//         if (map->map[i][j] != '1')
-//             print_error("Map not surrounded by walls");
-//         j++;
-//     }
-//     j = 0;
-//     while (map->map[i][j]) // last char in column
-//     {
-//         i = len_2d(map->map);
-//         while (map->map[i][j] == ' ')
-//         {
-//             check_inlets(map->map, i, j);
-//             i--;
-//         }
-//         if (map->map[i][j] != '1')
-//             print_error("Map not surrounded by walls");
-//         j++;
-//     }
-//     i = 0;
-//     while (map->map[i][j]) // first char in row
-//     {
-//         j = 0;
-//         while (map->map[i][j] == ' ')
-//         {
-//             check_inlets(map->map, i , j);
-//             j++;
-//         }
-//         if (map->map[i][j] != '1')
-//             print_error("Map not surrounded by walls");
-//         i++;
-//     }
-//     i = 0;
-//     while (map->map[i][j]) // last char in row
-//     {
-//         j = ft_strlen(map->map[i]) - 1;
-//         while (map->map[i][j] == ' ')
-//         {
-//             check_inlets(map->map, i, j);
-//             j--;
-//         }
-//         if (map->map[i][j] != '1')
-//             print_error("Map not surrounded by walls");
-//         i++;
-//     }
-// }
-
-static size_t	valid_position(t_map *map, size_t i, size_t j)
-{
-    if (!map->copy[i])
-        return (0);
-    else if (!map->copy[i][j])
-        return (0);
-	else if (map->copy[i][j] == '1')
-		return (0);
-    else if (map->copy[i][j] == 'a')
-		return (0);
-	return (1);
-}
-
-static void	fill_if_valid(t_map *map, int i, int j)
-{
-    if (i < 0 || j < 0)
-        return ;
-	else if (!valid_position(map, i, j))
-		return ;
-	map->copy[i][j] = 'a';
-	fill_if_valid(map, i - 1, j);
-	fill_if_valid(map, i + 1, j);
-	fill_if_valid(map, i, j - 1);
-	fill_if_valid(map, i, j + 1);
-}
-
-static void compare_maps(t_map *map)
-{
-    int i;
-    int j;
-
-    i = 0;
-    while (map->map[i])
-    {
-        j = 0;
-        while (map->map[i][j])
-        {
-            if (map->map[i][j] == ' ' && map->copy[i][j] == 'a')
-            {
-                free_2d(map->copy);
-                print_error("Map not surrounded by walls");
-            }
-            j++;
-        }
-        i++;
-    }
-}
-
-static void	check_walls(t_map *map)
-{
-	int     i;
-    int 	j;
-    int     flag;
-
-    i = 0;
-    flag = 0;
-    while (map->map[i])
-    {
-        j = 0;
-        while (map->map[i][j])
-        {
-            if (map->map[i][j] == 'N' || map->map[i][j] == 'S' || map->map[i][j] == 'E' || map->map[i][j] == 'W')
-            {
-                flag = 1;
-                break ;
-            }
-            j++;
-        }
-        if (flag == 1)
-            break ;
-        i++;
-    }
-	fill_if_valid(map, i, j);
-    compare_maps(map);
-}
-
 static void valid_chars(char c)
 {
     if (c != '1' && c != '0' && c != 'N' && c != 'S' && c != 'E' && c != 'W' && c != ' ')
@@ -286,28 +91,9 @@ static void find_empty_line(t_map *map)
     }
 }
 
-static void copy_map(t_map *map)
-{
-    int len;
-    int i;
-
-    len = 0;
-    len = len_2d(map->map);
-    map->copy = malloc(len * sizeof(char *) + 1);
-    i = 0;
-    while (map->map[i])
-    {
-        map->copy[i] = ft_strdup(map->map[i]);
-        i++;
-    }
-    map->copy[i] = NULL;
-}
-
 void    validate(t_map *map)
 {
     find_empty_line(map);
     check_chars(map);
-    copy_map(map);
     check_walls(map);
-    free_2d(map->copy);
 }
